Juani

+++ src/app/components/users/users.component.ts
@@ -7,6 +7,7 @@ import { Toast, ToastrService } from 'ngx-toastr';
 import { FilterPipeModule } from 'ngx-filter-pipe';
 import { NgxPaginationModule } from 'ngx-pagination';
 import { HttpErrorResponse } from '@angular/common/http';
+import { errorContext } from 'rxjs/internal/util/errorContext';
 
 @Component({
   selector: 'app-users',
@@ -35,16 +36,18 @@ export class UsersComponent implements OnInit {
   }
 
   deleteUser(id: number) {
-    this._userService.deleteUser(id).subscribe({
-      next: () => {
-        this.getUsers();
-        this.toastr.warning('User deleted successfully', 'User deleted');
-      },
-      error: (err: HttpErrorResponse) => {
-        this.toastr.error('El usuario tiene pedidos asociados', 'Error!');
-
-      }
-    });
+    if (confirm('¿Estás seguro de eliminar este usuario?')) {
+      this._userService.deleteUser(id).subscribe({
+        next: () => {
+          this.getUsers();
+          this.toastr.warning('Usuario eliminado correctamente', 'Usuario eliminado');
+        },
+        error: (err) => {
+          this.toastr.error(err.error.message, 'Error!');
+        }
+      });
+    }
+    
   }
 }
